requirements.txt
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
ffmpeg-python==0.2.0
Flask==3.1.0
Flask-SocketIO==5.5.1
future==1.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pycparser==2.22
pyheif==0.8.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
